Move Ares special game code to AresHandler.
