+++ src/ares/AresSpaceBonus.ts
@@ -1,0 +1,4 @@
+export enum AresSpaceBonus {
+  MC,
+  ANIMAL,
+}
