+++ src/cards/ares/CapitalAres.ts
@@ -1,6 +1,7 @@
 import { CardName } from "../../CardName";
 import { Capital } from "../Capital";
-import { AdjacencyBonus, AresSpaceBonus } from "../../ares/AdjacencyBonus";
+import { AdjacencyBonus } from "../../ares/AdjacencyBonus";
+import { AresSpaceBonus } from "../../ares/AresSpaceBonus";
 
 export class CapitalAres extends Capital {
   public name: CardName = CardName.CAPITAL_ARES;
