+++ src/cards/ares/CommercialDistrictAres.ts
@@ -1,6 +1,7 @@
 import { CardName } from "../../CardName";
 import { CommercialDistrict } from "../CommercialDistrict";
-import { AdjacencyBonus, AresSpaceBonus } from "../../ares/AdjacencyBonus";
+import { AdjacencyBonus } from "../../ares/AdjacencyBonus";
+import { AresSpaceBonus } from "../../ares/AresSpaceBonus";
 
 export class CommercialDistrictAres extends CommercialDistrict {
   public name: CardName = CardName.COMMERCIAL_DISTRICT_ARES;
