+++ src/cards/ares/EcologicalZoneAres.ts
@@ -1,6 +1,7 @@
 import { CardName } from "../../CardName";
 import { EcologicalZone } from "../EcologicalZone";
-import { AdjacencyBonus, AresSpaceBonus } from "../../ares/AdjacencyBonus";
+import { AdjacencyBonus } from "../../ares/AdjacencyBonus";
+import { AresSpaceBonus } from "../../ares/AresSpaceBonus";
 
 export class EcologicalZoneAres extends EcologicalZone {
   public name: CardName = CardName.ECOLOGICAL_ZONE_ARES;
