+++ src/cards/ares/NaturalPreserveAres.ts
@@ -1,6 +1,7 @@
 import { CardName } from "../../CardName";
 import { NaturalPreserve } from "../NaturalPreserve";
-import { AdjacencyBonus, AresSpaceBonus } from "../../ares/AdjacencyBonus";
+import { AdjacencyBonus } from "../../ares/AdjacencyBonus";
+import { AresSpaceBonus } from "../../ares/AresSpaceBonus";
 
 export class NaturalPreserveAres extends NaturalPreserve {
   public name: CardName = CardName.NATURAL_PRESERVE_ARES;
